Quash __int64 to int conversion warnings, by replacing int with ptrdiff_t, long with size_t, streamsize etc, as appropriate
SVN-Revision: 27620

+++ core/vsl/tests/test_vlarge_block_io.cxx
@@ -59,7 +59,7 @@ void test_vlarge_block(void * block, vcl_size_t s, T scale)
     if (numbers[i] != static_cast<T>(vcl_numeric_limits<T>::max() - T(i)) )
       errors++;
 
-  TEST_NEAR("No errors in stored numbers", errors, 0, 0);
+  TEST_NEAR("No errors in stored numbers", static_cast<double>(errors), 0, 0);
 }
